refactor capnp compare
