+++ bfasst/tools/compare/structural/phys_capnp_rules.ninja.mustache
@@ -1,0 +1,4 @@
+rule phys_capnp_cmp
+    command = python {{ bfasst_path }}/bfasst/utils/structural_capnp.py --build_dir $build_dir --impl_dcp $impl_dcp --impl_edf $impl_edf --phys_capnp $phys_capnp --edf_capnp $edf_capnp {{#log_name}}--log_name $log_name {{/log_name}}--logging_level $logging_level
+    description = generate physical netlist for $build_dir
+
