+++ bfasst/tools/transform/phys_netlist.py
@@ -15,33 +15,17 @@ from bfasst.utils import compare_json
 class PhysNetlist(Tool):
     """Create rule and build snippets for phys netlist creation."""
 
-    def __init__(
-        self,
-        flow,
-        design,
-        impl_checkpoint,
-        impl_edf,
-        *,
-        phys_capnp=None,
-        edf_capnp=None,
-        logging_level="INFO",
-    ):
+    def __init__(self, flow, design, impl_checkpoint, impl_edf, *, logging_level="INFO"):
         super().__init__(flow, design)
 
         self.impl_checkpoint = impl_checkpoint
         self.impl_edf = impl_edf
-        self.phys_capnp = phys_capnp
-        self.edf_capnp = edf_capnp
         self.logging_level = logging_level
 
         self.build_path = self.design_build_path / "vivado_phys_netlist"
 
         self._init_outputs()
         self.rule_snippet_path = NINJA_TRANSFORM_TOOLS_PATH / "phys_netlist_rules.ninja.mustache"
-        self.rules_render_dict = {
-            "phys_capnp": self.phys_capnp,
-            "edf_capnp": self.edf_capnp,
-        }
 
     def create_build_snippets(self):
         self.__write_json_file()
@@ -63,35 +47,18 @@ class PhysNetlist(Tool):
                 f.write(checkpoint_to_v_json)
 
     def __append_build_snippets(self):
-        # the phys netlist tool can be used in 2 different ways so we need a way to differentiate
-        if self.flow.__class__.__name__ in {"VivadoPhysCapnp"}:
-            with open(NINJA_TRANSFORM_TOOLS_PATH / "phys_netlist_capnp_build.ninja.mustache") as f:
-                phys_netlist_ninja = chevron.render(
-                    f,
-                    {
-                        "phys_netlist_output": self.build_path,
-                        "phys_netlist_library": NINJA_TRANSFORM_TOOLS_PATH,
-                        "build_dir": self.build_path.parent,
-                        "logging_level": self.logging_level,
-                        "impl_dcp": self.impl_checkpoint,
-                        "impl_edf": self.impl_edf,
-                        "phys_capnp": self.phys_capnp,
-                        "edf_capnp": self.edf_capnp,
-                    },
-                )
-        else:
-            with open(NINJA_TRANSFORM_TOOLS_PATH / "phys_netlist_build.ninja.mustache") as f:
-                phys_netlist_ninja = chevron.render(
-                    f,
-                    {
-                        "phys_netlist_output": self.build_path,
-                        "phys_netlist_library": NINJA_TRANSFORM_TOOLS_PATH,
-                        "build_dir": self.build_path.parent,
-                        "logging_level": self.logging_level,
-                        "impl_dcp": self.impl_checkpoint,
-                        "impl_edf": self.impl_edf,
-                    },
-                )
+        with open(NINJA_TRANSFORM_TOOLS_PATH / "phys_netlist_build.ninja.mustache") as f:
+            phys_netlist_ninja = chevron.render(
+                f,
+                {
+                    "phys_netlist_output": self.build_path,
+                    "phys_netlist_library": NINJA_TRANSFORM_TOOLS_PATH,
+                    "build_dir": self.build_path.parent,
+                    "logging_level": f"--logging_level {self.logging_level}",
+                    "impl_dcp": self.impl_checkpoint,
+                    "impl_edf": self.impl_edf,
+                },
+            )
 
         with open(NINJA_BUILD_PATH, "a") as f:
             f.write(phys_netlist_ninja)
@@ -109,5 +76,8 @@ class PhysNetlist(Tool):
     def add_ninja_deps(self, deps):
         self._add_ninja_deps_default(deps, __file__)
         deps.append(BFASST_UTILS_PATH / "rw_phys_netlist.py")
+        deps.append(BFASST_UTILS_PATH / "general.py")
         deps.append(BFASST_UTILS_PATH / "rw_helpers.py")
+        deps.append(NINJA_TRANSFORM_TOOLS_PATH / "phys_netlist_rules.ninja.mustache")
+        deps.append(NINJA_TRANSFORM_TOOLS_PATH / "phys_netlist_build.ninja.mustache")
         deps.append(NINJA_TRANSFORM_TOOLS_PATH / "checkpoint_to_v.tcl.mustache")
