--- bfasst/tools/transform/phys_netlist_capnp_build.ninja.mustache
@@ -1,12 +1,0 @@
-build {{ phys_netlist_output }}/viv_impl_physical.edf {{ phys_netlist_output }}/phys_netlist.dcp {{ phys_netlist_output }}/{{ log_name }} {{ phys_netlist_output }}/rapidwright_stdout.log: phys_netlist | bfasst/utils/rw_phys_netlist.py {{ impl_edf }} {{ impl_dcp }} {{ phys_capnp }} {{ edf_capnp }}
-    logging_level = {{ logging_level }}
-    impl_dcp = {{ impl_dcp }}
-    impl_edf = {{ impl_edf }}
-    log_name = {{ log_name }}
-    {{#phys_capnp}}
-    phys_capnp = {{ . }} 
-    {{/phys_capnp}}
-    {{#edf_capnp}}
-    edf_capnp = {{ . }} 
-    {{/edf_capnp}}
-    build_dir = {{ build_dir }}
