+++ bfasst/utils/rw_phys_netlist.py
@@ -5,24 +5,16 @@
 
 from argparse import ArgumentParser
 import logging
-from itertools import groupby
 from pathlib import Path
 import sys
 import time
-from typing import Optional
-
-# import code
-# import traceback
 
 import jpype
 from jpype.types import JInt
 
 from bfasst import jpype_jvm
 from bfasst.config import PART
-from bfasst.utils import convert_verilog_literal_to_int
-from bfasst.utils.capnp_cells import CapnpCells
 import bfasst.utils.rw_helpers as rw
-from bfasst.utils.structural_helpers import create_cell_props
 
 
 # pylint: disable=wrong-import-position,wrong-import-order,import-error
@@ -33,15 +25,11 @@ from com.xilinx.rapidwright.edif import (
     EDIFNet,
     EDIFCell,
     EDIFCellInst,
-    EDIFHierPortInst,
-    EDIFHierNet,
     EDIFHierCellInst,
     EDIFPropertyValue,
     EDIFValueType,
 )
 from com.xilinx.rapidwright.interchange import (
-    LogNetlistReader,
-    PhysNetlistReader,
     LogNetlistWriter,
     PhysNetlistWriter,
 )
@@ -55,10 +43,6 @@ class PhysNetlistTransformError(Exception):
     """Exception for errors in phys netlist transformation"""
 
 
-class StructuralCompareError(Exception):
-    """Exception for structural comparison errors"""
-
-
 class RwPhysNetlist:
     """Creates a xilinx netlist that has only physical primitives"""
 
@@ -67,33 +51,21 @@ class RwPhysNetlist:
         self.stage_dir = self.build_dir / "vivado_phys_netlist"
         (self.stage_dir / log_name).unlink(missing_ok=True)
         self.logging_level = logging_level
+        log_path = self.stage_dir / log_name if "/" not in log_name else Path(log_name)
         logging.basicConfig(
-            filename=self.stage_dir / log_name,
+            filename=log_path,
             format="%(asctime)s %(message)s",
             level=self.logging_level,
             datefmt="%Y%m%d%H%M%S",
         )
+        # Redirect rapidwright output to file
+        rapidwright_log_path = str(self.stage_dir / "rapidwright_stdout.log")
+        System.setOut(PrintStream(File(rapidwright_log_path)))
+        System.setErr(PrintStream(File(rapidwright_log_path)))
 
         # Rapidwright design / netlist
         self.vivado_design = None
         self.vivado_netlist = None
-        self.rev_design = None
-        self.rev_netlist = None
-
-        self.capnp_cells = None
-        self.rw_value_mismatch = 0
-        self.rw_value_mismatches = []
-        self.rw_problem_cells = set()
-
-        self.rw_port_mismatch = 0
-        self.rw_port_mismatch_cells = set()
-
-        self._cell_props = create_cell_props()
-        self.matches = {}  # vivado edif cell name: (vivado_edif_cell, rev_edif_cell)
-        # nets are named based on the net driver.
-        self.net_map = {}  # net names: [net names and alias nets] (vivado: rev)
-        self.driver_cache = {}  # EDIFNet: driver_full_name_str
-        self.rev_driver_cache = {}  # EDIFHierPortInst: driver_full_name_str
 
         # Const nets
         self.vcc = None
@@ -109,6 +81,7 @@ class RwPhysNetlist:
 
         # Keep a list of old replaced cells to remove after processing
         self.cells_to_remove = []
+        self.phys_cells = []  # tuple(EDIFCellInst, SiteInst, str(bel_name))
 
         self.lut_pair_bel_names = [
             ("A6LUT", "A6LUT_O6", "A5LUT", "A5LUT_O5"),
@@ -117,74 +90,10 @@ class RwPhysNetlist:
             ("D6LUT", "D6LUT_O6", "D5LUT", "D5LUT_O5"),
         ]
 
-    def get_properties_for_type(self, cell_type: str) -> tuple[str]:
-        """Return the list of properties that must match for a given cell type
-        for the cell to be considered equivalent."""
-        try:
-            return self._cell_props[cell_type]
-        except KeyError as err:
-            raise StructuralCompareError(f"Unhandled properties for type {cell_type}") from err
-
-    def run(
-        self,
-        impl_netlist_checkpoint: Path,
-        impl_netlist_edif_path: Path,
-        phys_capnp: Optional[Path] = None,
-        edf_capnp: Optional[Path] = None,
-    ) -> None:
+    def run(self, impl_netlist_checkpoint: Path, impl_netlist_edif_path: Path) -> None:
         """Transform the logical netlist into a netlist with only physical primitives"""
-        phys_netlist_edif_path = self.stage_dir / "viv_impl_physical.edf"
-
-        # Redirect rapidwright output to file
-        rapidwright_log_path = str(self.stage_dir / "rapidwright_stdout.log")
-        System.setOut(PrintStream(File(rapidwright_log_path)))
-        System.setErr(PrintStream(File(rapidwright_log_path)))
-
-        phys_netlist_checkpoint = self.stage_dir / "phys_netlist.dcp"
-        logging.info(
-            "Starting logical to physical netlist conversion for %s %s",
-            f"{self.build_dir / 'in_context' / 'impl' / 'impl.dcp'}",
-            f"{self.build_dir / 'in_context' / 'impl' / 'viv_impl.edf'}",
-        )
-
-        start_time = time.time()
-        # Catch all Java exceptions since they are not picklable,
-        # and so cannot be handled properly by multiprocessing
-        # Don't raise from as this is also problematic.
-        try:
-            self.__run_rapidwright(
-                impl_netlist_checkpoint,
-                impl_netlist_edif_path,
-                phys_capnp,
-                edf_capnp,
-            )
-        except jpype.JException as exc:  # pylint: disable=no-member
-            raise rw.RapidwrightException from exc  # pylint: disable=bad-exception-cause
-        end_time = time.time()
-
-        # Export checkpoint, then run vivado to generate a new netlist
-        self.vivado_design.unplaceDesign()
-        self.vivado_design.writeCheckpoint(phys_netlist_checkpoint)
-
-        logging.info("Writing EDIF phsyical netlist: %s", phys_netlist_edif_path)
-        self.vivado_netlist.exportEDIF(phys_netlist_edif_path)
-        logging.info(
-            "Writing capnp interchange netlist: %s",
-            str(self.stage_dir / "phys_logical_netlist.capnp"),
-        )
-        LogNetlistWriter.writeLogNetlist(
-            self.vivado_netlist, str(self.stage_dir / "phys_logical_netlist.capnp")
-        )
-        logging.info(
-            "Writing capnp interchange physical netlist: %s",
-            str(self.stage_dir / "phys_physical_netlist.capnp"),
-        )
-        PhysNetlistWriter.writePhysNetlist(
-            self.vivado_design, str(self.stage_dir / "phys_physical_netlist.capnp")
-        )
-        logging.info("Transformation time %s seconds", f"{end_time - start_time:0.2f}")
-        with open(self.stage_dir / "transformation_time.txt", "w") as fp:
-            fp.write(f"{end_time - start_time:.2f}\n")
+        self.run_rapidwright(impl_netlist_checkpoint, impl_netlist_edif_path)
+        self.export_transformation()
 
     def __init_const_nets(self) -> None:
         """Init VCC and GND nets"""
@@ -221,41 +130,49 @@ class RwPhysNetlist:
             assert const_port
             const_net.createPortInst(const_port, const_edif_inst)
 
-    def load_designs(
-        self, impl_dcp: Path, impl_edf: Path, phys_capnp: Path, edf_capnp: Path
-    ) -> None:
+    def load_designs(self, impl_dcp: Path, impl_edf: Path) -> None:
         """Load the designs from the given paths"""
-        device = Device.getDevice(PART)
         logging.info("Loading vivado dcp and edf files: %s, %s", str(impl_dcp), str(impl_edf))
         start_time = time.time()
         self.vivado_design = Design.readCheckpoint(impl_dcp, impl_edf)
         self.vivado_netlist = self.vivado_design.getNetlist()
         self.vivado_design.flattenDesign()
-        self.vivado_netlist.expandMacroUnisims(device.getSeries())
+        self.vivado_netlist.expandMacroUnisims(Device.getDevice(PART).getSeries())
         logging.info("Loading vivado netlist took %s seconds.", time.time() - start_time)
-        if phys_capnp is not None:
-            assert edf_capnp is not None
-            logging.info("Loading reversed capnp objects: %s, %s", str(phys_capnp), str(edf_capnp))
-            start_time = time.time()
-            self.rev_netlist = LogNetlistReader.readLogNetlist(str(edf_capnp))
-            self.rev_design = PhysNetlistReader.readPhysNetlist(str(phys_capnp), self.rev_netlist)
-            self.rev_design.flattenDesign()
-            self.rev_netlist.expandMacroUnisims(device.getSeries())
-            logging.info(
-                "Loading reversed capnp objects took %s seconds.", time.time() - start_time
-            )
 
-            self.capnp_cells = CapnpCells(phys_capnp, edf_capnp)
-        else:
-            assert edf_capnp is None
-            logging.info("No capnp objects provided, skipping comparison")
-            self._compare_cell = self.__dummy_compare
+    def export_transformation(self):
+        """Export checkpoint, then run vivado to generate a new netlist"""
+        phys_netlist_checkpoint = self.stage_dir / "phys_netlist.dcp"
+        self.vivado_design.unplaceDesign()
+        self.vivado_design.writeCheckpoint(phys_netlist_checkpoint)
 
-    def __run_rapidwright(
-        self, impl_dcp: Path, impl_edf: Path, phys_capnp: Path, edf_capnp: Path
-    ) -> None:
+        phys_netlist_edif_path = self.stage_dir / "viv_impl_physical.edf"
+        logging.info("Writing EDIF phsyical netlist: %s", phys_netlist_edif_path)
+        self.vivado_netlist.exportEDIF(phys_netlist_edif_path)
+        logging.info(
+            "Writing capnp interchange netlist: %s",
+            str(self.stage_dir / "phys_logical_netlist.capnp"),
+        )
+        LogNetlistWriter.writeLogNetlist(
+            self.vivado_netlist, str(self.stage_dir / "phys_logical_netlist.capnp")
+        )
+        logging.info(
+            "Writing capnp interchange physical netlist: %s",
+            str(self.stage_dir / "phys_physical_netlist.capnp"),
+        )
+        PhysNetlistWriter.writePhysNetlist(
+            self.vivado_design, str(self.stage_dir / "phys_physical_netlist.capnp")
+        )
+
+    def run_rapidwright(self, impl_dcp: Path, impl_edf: Path) -> None:
         """Do all rapidwright related processing on the netlist"""
-        self.load_designs(impl_dcp, impl_edf, phys_capnp, edf_capnp)
+        logging.info(
+            "Starting logical to physical netlist conversion for %s %s",
+            str(impl_dcp),
+            str(impl_edf),
+        )
+        start_time = time.time()
+        self.load_designs(impl_dcp, impl_edf)
 
         self.__init_const_nets()
 
@@ -321,9 +238,7 @@ class RwPhysNetlist:
                 "LDCE",
                 "DSP48E1",
             ):
-                self._compare_cell(
-                    edif_cell_inst, cell.getName(), cell.getSiteInst(), cell.getBELName()
-                )
+                self.phys_cells.append((edif_cell_inst, cell.getSiteInst(), cell.getBELName()))
                 continue
 
             logging.error("Unsupported cell type %s. Exiting...", cell_type)
@@ -331,21 +246,16 @@ class RwPhysNetlist:
 
         logging.info("Rapidwright bug count %d", self.rw_value_mismatch)
         logging.info("Processed other cells in %s seconds", time.time() - start_time)
-        # print("Processed other cells in %s seconds", time.time() - start_time)
-        # print(f"Cmp cell time: %s", self.cmp_cell_time)
-        # exit()
-        if edf_capnp is not None:
-            self.fix_rev_dsps()
-            start_time = time.time()
-            # Check nets of matched cells
-            self._check_nets()
-            logging.info("Checked nets in %s seconds", time.time() - start_time)
-            # print("Checked nets in %s seconds", time.time() - start_time)
 
         # Remove old unusued cells
         logging.info("Removing old cells...")
         for cell in self.cells_to_remove:
             rw.remove_and_disconnect_cell(cell)
+
+        ttime = time.time() - start_time
+        logging.info("Transformation time %s seconds", f"{ttime:0.2f}")
+        with open(self.stage_dir / "transformation_time.txt", "w") as fp:
+            fp.write(f"{ttime:.2f}\n")
 
     def __process_all_luts(self, cells_already_visited: set[Cell]) -> None:
         """Visit all LUTs and replace them with LUT6_2 instances"""
@@ -464,7 +374,7 @@ class RwPhysNetlist:
             if lut6[1] in const_net:
                 # If a gnd net, then there can't be a cell there
                 assert lut6[0] is None
-                if lut5[0] is not None:
+                if lut5[0] is not None and "LUT" in lut5[0].getType():
                     self.__process_lut5_and_const_lut(lut5[0], lut6[1], site_inst, is_gnd)
                     return {
                         lut5[0],
@@ -477,7 +387,7 @@ class RwPhysNetlist:
                 # If a gnd net, then there can't be a cell there
                 # This assumption is not true for LUTRAMs
                 assert lut5[0] is None
-                if lut6[0] is not None:
+                if lut6[0] is not None and "LUT" in lut6[0].getType():
                     self.__process_lut5_and_const_lut(lut6[0], lut5[1], site_inst, is_gnd)
                     return {
                         lut6[0],
@@ -681,9 +591,7 @@ class RwPhysNetlist:
         """
 
         if rw.PinMap.cell_is_default_mapping(cell):
-            self._compare_cell(
-                cell.getEDIFCellInst(), cell.getName(), cell.getSiteInst(), cell.getBELName()
-            )
+            self.phys_cells.append((cell.getEDIFCellInst(), cell.getSiteInst(), cell.getBELName()))
             return []
 
         logging.info("\tInputs are unexpectedly permuted... Exiting")
@@ -696,9 +604,7 @@ class RwPhysNetlist:
         raise NotImplementedError if they are.
         """
         if rw.PinMap.cell_is_default_mapping(cell):
-            self._compare_cell(
-                cell.getEDIFCellInst(), cell.getName(), cell.getSiteInst(), cell.getBELName()
-            )
+            self.phys_cells.append((cell.getEDIFCellInst(), cell.getSiteInst(), cell.getBELName()))
             return []
 
         logging.info("\tInputs are permuted. Ensuring connections")
@@ -748,8 +654,7 @@ class RwPhysNetlist:
             port = bufgctrl.getPort(port_name)
             assert port
             self.vcc.createPortInst(port, bufgctrl)
-
-        self._compare_cell(bufgctrl, bufgctrl.getName(), bufg_cell.getSiteInst(), "BUFGCTRL")
+        self.phys_cells.append((bufgctrl, bufg_cell.getSiteInst(), "BUFGCTRL"))
 
         return [bufg_cell]
 
@@ -872,8 +777,7 @@ class RwPhysNetlist:
         for logical_port in rw.PinMap["LUT6_2"]:
             if logical_port.startswith("I"):
                 self.vcc.createPortInst(new_cell_inst.getPort(logical_port), new_cell_inst)
-
-        self._compare_cell(new_cell_inst, new_cell_inst.getName(), site_inst, f"{pins[0][0]}6LUT")
+        self.phys_cells.append((new_cell_inst, site_inst, f"{pins[0][0]}6LUT"))
 
     def __process_lut(
         self, lut6_cell: Cell, lut5_cell: Cell, lut5_only: bool = False
@@ -968,9 +872,7 @@ class RwPhysNetlist:
         logging.info("  New LUT INIT: %s", new_init)
         new_cell_inst.addProperty("INIT", new_init)
 
-        self._compare_cell(
-            new_cell_inst, lut6_cell.getName(), lut6_cell.getSiteInst(), lut6_cell.getBELName()
-        )
+        self.phys_cells.append((new_cell_inst, lut6_cell.getSiteInst(), lut6_cell.getBELName()))
 
         return new_cell_inst
 
@@ -1037,7 +939,7 @@ class RwPhysNetlist:
 
         # Fix the new LUT INIT property based on the new pin mappings
         rw.process_shared_gnd_lut_eqn(lut5, const_pin, new_cell_inst, is_gnd)
-        self._compare_cell(new_cell_inst, lut5.getName(), site_inst, lut5.getBELName())
+        self.phys_cells.append((new_cell_inst, site_inst, lut5.getBELName()))
         return lut5
 
     def __create_lut_routethru_net(
@@ -1101,286 +1003,6 @@ class RwPhysNetlist:
         else:
             new_net.createPortInst(routed_to_port_inst.getPort(), routed_to_cell_inst)
 
-    def handle_missing_prop(self, rev_cell: Cell, prop_name: str) -> int:
-        """
-        Sometimes properties from capnp are missing in the RapidWright data structure.
-        So far this has only been LUT Init strings.
-        """
-        _, lcapnp_cell = self.capnp_cells.get_capnp_cell(rev_cell.getName())
-        assert lcapnp_cell is not None
-        rev_value = None
-        for props in lcapnp_cell.propMap.entries:
-            if self.capnp_cells.log_capnp.strList[props.key] == prop_name:
-                rev_value = self.capnp_cells.log_capnp.strList[props.textValue]
-                break
-        assert rev_value is not None
-        rev_value = convert_verilog_literal_to_int(rev_value)
-        self.rw_value_mismatch += 1
-        self.rw_problem_cells.add(rev_cell.getName())
-        logging.warning("Adding cell %s to rw problem cells", rev_cell.getName())
-        return rev_value
-
-    def __dummy_compare(self, ecell, log_name, site, bel_name):
-        pass
-
-    # pylint: disable=method-hidden
-    def _compare_cell(
-        self, ecell: EDIFCellInst, log_name: str, site: SiteInst, bel_name: str
-    ) -> None:
-        """
-        Compare the post-implementation cell to the reversed cell.
-
-        ecell: The transformed edif cell (used for properties and connections)
-        log_name: Logical name of the cell
-        cell: The original post-implementation cell (used for location data)
-        """
-
-        rev_cell = self.rev_design.getSiteInst(site.getName()).getCell(bel_name)
-
-        logging.info(
-            "Comparing cell %s on BEL %s to reversed cell %s",
-            log_name,
-            bel_name,
-            rev_cell.getName(),
-        )
-
-        rev_ecell = rev_cell.getEDIFCellInst()
-
-        if ecell.getCellType().getName() != rev_ecell.getCellType().getName():
-            if "LUT" not in ecell.getCellType().getName():
-                raise StructuralCompareError(
-                    f"Cell {ecell.getCellType().getName()} does not match reversed cell "
-                    + f"{rev_ecell.getCellType().getName()}."
-                )
-            if __debug__:
-                # Sanity checks when LUT6_2 is split into LUT6 and LUT5
-                logging.warning(
-                    "Warning: Comparing LUT cell %s to %s",
-                    ecell.getCellType().getName(),
-                    rev_ecell.getCellType().getName(),
-                )
-
-        # Check properties
-        cell_props = ecell.getPropertiesMap()  # The regular cell may have out of date properties
-        rev_props = rev_cell.getProperties()
-        keys = self.get_properties_for_type(ecell.getCellType().getName())
-
-        for name in keys:
-            if name not in rev_props:
-                raise StructuralCompareError(
-                    f"Property {name} not in rev cell {rev_cell.getName()} properties."
-                )
-            value = convert_verilog_literal_to_int(cell_props[name].getValue())
-            rev_value = convert_verilog_literal_to_int(rev_props[name].getValue())
-
-            if rev_value != value:
-                capnp_rev_value = self.handle_missing_prop(rev_cell, name)
-                if capnp_rev_value != value:
-                    raise StructuralCompareError(
-                        f"Property {name} in rev cell {rev_cell.getName()} does not match. "
-                        + f"({value} != {capnp_rev_value} != {rev_value}"
-                    )
-
-        self.matches[ecell.getName()] = (ecell, rev_cell, rev_ecell)
-
-    # pylint: enable=method-hidden
-
-    def _compare_shared_lut_pins(
-        self, rev_port_insts: dict[str, EDIFHierPortInst], other_lut: EDIFHierCellInst
-    ) -> None:
-        """
-        Compare the shared pins of LUT6 and LUT5 cells to ensure they match.
-        This is used when a LUT6_2 cell is split into a LUT6 and LUT5.
-        """
-        for port_name in ("I0", "I1", "I2", "I3", "I4"):
-            if port_name not in rev_port_insts:
-                assert other_lut.getPortInst(port_name) is None
-                continue
-            l6_net = rev_port_insts[port_name].getHierarchicalNet()
-            l5_net = other_lut.getPortInst(port_name).getHierarchicalNet()
-            assert l6_net == l5_net
-
-    def update_lut_ports(self, rev_cell: Cell, rev_port_insts: dict[str, EDIFHierPortInst]) -> None:
-        """Update port_insts on the reversed LUT to reflect the combined LUT6_2 cell."""
-        assert "O6" not in rev_port_insts
-        assert "O5" not in rev_port_insts
-
-        if rev_cell.getType() == "LUT6":
-            rev_port_insts["O6"] = rev_port_insts.pop("O")
-            rev_lut5 = rev_cell.getSiteInst().getCell(f"{rev_cell.getBELName()[0]}5LUT")
-            if rev_lut5 is not None:
-                rev_hecell_lut5 = rev_lut5.getEDIFHierCellInst()
-                o5 = rev_hecell_lut5.getPortInst("O")
-                if o5 is not None:
-                    rev_port_insts["O5"] = o5
-                if __debug__:  # Sanity check input nets are the same on shared pins
-                    self._compare_shared_lut_pins(rev_port_insts, rev_hecell_lut5)
-
-        elif rev_cell.getType() == "LUT5":
-            rev_port_insts["O5"] = rev_port_insts.pop("O")
-            rev_lut6 = rev_cell.getSiteInst().getCell(f"{rev_cell.getBELName()[0]}6LUT")
-            if rev_lut6 is not None:
-                rev_hecell_lut6 = rev_lut6.getEDIFHierCellInst()
-                rev_port_insts["O6"] = rev_hecell_lut6.getPortInst("O")
-                rev_port_insts["I5"] = rev_hecell_lut6.getPortInst("I5")
-                if __debug__:  # Sanity check input nets are the same on shared pins
-                    self._compare_shared_lut_pins(rev_port_insts, rev_hecell_lut6)
-
-    def _check_nets(self) -> None:
-        """
-        Check nets at the end once all transformations are done.
-        Assume nets match -> throw an error if an already matched net is contradicted
-        """
-        # Loop through each pair of matching cells
-        for name, (ecell, rev_cell, _) in self.matches.items():
-            ecell_port_insts = ecell.getPortInsts()
-            rev_hecell = rev_cell.getEDIFHierCellInst()
-            rev_port_insts = {p.getPortInst().getName(): p for p in rev_hecell.getHierPortInsts()}
-
-            # Check each cell has the same number of connections
-            if ecell.getCellType().getName() == "LUT6_2":
-                self.update_lut_ports(rev_cell, rev_port_insts)
-
-            # Get the corresponding ports
-            for port in ecell_port_insts:
-                # Get the driver for the net
-                net = port.getNet()
-                if net not in self.driver_cache:
-                    self.driver_cache[net] = net.getSourcePortInsts(True)[0].getFullName()
-                net_driver = self.driver_cache[net]
-
-                # Get the driver for the reversed net
-                rev_port = rev_port_insts.pop(port.getName(), None)
-                rev_net = rev_port.getHierarchicalNet() if rev_port is not None else None
-                if rev_net is None:
-                    # Assume ground connection
-                    rev_net = self.rev_design.getGndNet()
-                    rev_net = self.rev_netlist.getHierNetFromName(rev_net.getName())
-                rev_net_driver = self._get_single_rev_driver(rev_port, rev_net, net_driver)
-
-                # If the net is not in the mapping, add it
-                if net_driver not in self.net_map:
-                    logging.info("Mapping net %s to %s", net_driver, rev_net_driver)
-                    self.net_map[net_driver] = rev_net_driver
-
-                # If the net is already mapped to a different rev net, raise an error
-                elif self.net_map[net_driver] != rev_net_driver:
-                    if net.isGND() and not rw.is_connected(rev_port):
-                        # F2B port is connected to a net that does not have a real driver (GND).
-                        continue
-
-                    # check for routethru or const generator
-
-                    if rev_port is None:
-                        rev_port = (
-                            f"{rev_hecell.getFullHierarchicalInstName()} "
-                            + "with Disconnected port... Assuming ground connection"
-                        )
-                    # self.rw_port_mismatch += 1
-                    # self.rw_port_mismatch_cells.add(rev_hecell.getFullHierarchicalInstName())
-                    # logging.shutdown()
-                    # # traceback.print_exc()
-                    # p = self.capnp_cells.phys_capnp
-                    # n = self.capnp_cells.log_capnp
-                    # capnp_cell, lcapnp_cell = self.capnp_cells.get_capnp_cell(rev_cell.getName())
-                    # import code
-
-                    # code.interact(local=dict(globals(), **locals()))
-                    raise StructuralCompareError(
-                        f"Net {net_driver} on port {port.getName()} in cell {ecell.getName()}"
-                        + f" already mapped. ({self.net_map[net_driver]} != {rev_net_driver})\n\t"
-                        + f"Edif Cell/port: {name}/{port.getName()}  \n\t"
-                        + f"Rev Edif Cell/port: {str(rev_port)}"
-                    )
-            for port in rev_port_insts.values():
-                # These ports are disconnected in the Vivado design
-                if rw.is_connected(port):
-                    raise StructuralCompareError(
-                        f"Reverse Port {port} in cell {name} is "
-                        + "connected, but disconnected in Vivado"
-                    )
-
-    def _get_single_rev_driver(
-        self, rev_port: EDIFHierPortInst, rev_net: EDIFHierNet, driver: str
-    ) -> str:
-        """
-        Helper to return exactly one driver-string for rev_net,
-        caching results and handling the 2-driver IBUF case.
-        """
-        if rev_port not in self.rev_driver_cache:
-            leafs = rev_net.getLeafHierPortInsts(True, True)
-            drivers = []
-            sinks = []
-            for k, g in groupby(leafs, key=lambda x: x.isOutput()):
-                if k:
-                    drivers.extend(g)
-                else:
-                    sinks.extend(g)
-            if not drivers:
-                # fallback for top-level I/Os
-                drivers = [p for p in rev_net.getPortInsts() if p.isInput()]
-                if len(drivers) == 2:
-                    # IBUF nets: pick the one that matches driver
-                    drivers = [l for l in drivers if str(l) == driver]
-            assert len(drivers) == 1, f"Expected 1 driver on {rev_net}, found {drivers}"
-            tmp = dict.fromkeys(sinks, str(drivers[0]))
-            self.rev_driver_cache.update(tmp)
-            self.rev_driver_cache[drivers[0]] = str(drivers[0])
-            self.rev_driver_cache[rev_port] = str(drivers[0])
-
-        return self.rev_driver_cache[rev_port]
-
-    def fix_rev_dsps(self) -> None:
-        """
-        Fasm2bels does not accurately set the OPMODE, ALUMODE, INMODE pins.
-        The fasm file specifies some incoming nets to these ports should be
-        inverted, but the f2b code does not do anything with this information.
-        I have added some code to add a custom property to the DSP48E2 for these
-        values. Check the dsp properties and adjust the input signals accordingly.
-        """
-        props = ("IS_INMODE_INVERTED", "IS_ALUMODE_INVERTED", "IS_OPMODE_INVERTED")
-        constant_ports = ["CARRYCASCIN", "CARRYCASCOUT"] + [
-            f"{port_name}[{idx}]"
-            for port_name, bus_size in (("PCIN", 48), ("PCOUT", 48))
-            for idx in range(bus_size)
-        ]
-        for cell in self.rev_design.getCells():
-            if cell.getType() != "DSP48E1":
-                continue
-            logging.info("Processing reversed DSP %s for inverted signals", cell.getName())
-            ecell = cell.getEDIFCellInst()
-            assert ecell
-
-            # XRAY erroneously does not output feature to invert carryin signal.
-            rw.flip_const_port_signal(self.rev_design, ecell, "CARRYIN")
-
-            # These signals are fixed routing paths in the FPGA (nonprogrammable). They are always
-            # connected to the cooresponding neighbor DSP site in the tile. Internal to the site,
-            # there is a mux that selects them. Dispite this, when they are unused, Vivado shows
-            # them as connected to GND or simply disconnected. F2B shows the persistant connection
-            # to the neighboring DSP site.
-            # Other inputs control the mux, so as long as this matches, the cascade values
-            # are irrelevant. See Xilinx UG479 Figure 2-1 for more information.
-            site_name = cell.getSite().getName()
-            viv_cell = self.vivado_design.getSiteInst(site_name).getCell(cell.getBELName())
-            viv_ecell = viv_cell.getEDIFCellInst()
-            for port_name in constant_ports:
-                viv_port_inst = viv_ecell.getPortInst(port_name)
-                if viv_port_inst is None or viv_port_inst.getNet().isGND():
-                    logging.info("Disconnecting %s port", port_name)
-                    port_inst = ecell.getPortInst(port_name)
-                    assert port_inst
-                    port_inst.getNet().removePortInst(port_inst)
-
-            # Flip dsp signals according to f2b properties
-            for prop in props:
-                val = ecell.getProperty(prop).getValue()
-                port_name = prop.split("_")[1]
-                for idx, inv in enumerate(reversed(val[3:])):
-                    if not int(inv):
-                        continue
-                    rw.flip_const_port_signal(self.rev_design, ecell, f"{port_name}[{idx}]", idx)
-
 
 if __name__ == "__main__":
     parser = ArgumentParser()
@@ -1403,20 +1025,6 @@ if __name__ == "__main__":
         help="The implementation edf file to use for the netlist.",
     )
     parser.add_argument(
-        "--phys_capnp",
-        type=Path,
-        required=False,
-        default=None,
-        help="The capnp physical netlist to compare against",
-    )
-    parser.add_argument(
-        "--edf_capnp",
-        type=Path,
-        required=False,
-        default=None,
-        help="The capnp logical netlist to compare against",
-    )
-    parser.add_argument(
         "--logging_level", default="INFO", help="Decides what levels of logs to display"
     )
     parser.add_argument(
@@ -1426,6 +1034,6 @@ if __name__ == "__main__":
     netlist_generator = RwPhysNetlist(args.build_dir, args.logging_level, args.log_name)
     try:
         netlist_generator.run(args.impl_dcp, args.impl_edf, args.phys_capnp, args.edf_capnp)
-    except jpype.JException as e:  # pylint: disable=no-member
+    except (jpype.JException, rw.RapidwrightException) as e:  # pylint: disable=no-member
         logging.error("ERROR: %s", e)
         sys.exit(1)
